update use of autotools
There is quite a lot going on in this change.  The sort summary is: this is
a build system modernization.  Needless to say there is no way to do this
kind of update without causing some compatibility issues, so lets hope they
do not bother users too much.  Various changes includes:

* Generated files are removed from version control.  When building from git
  clone one has to first run ./autogen.sh
* Use pkg-config when ever possible.
* --with-logfile=PATH and --with-pidfile=PATH defaults include prefix.
* Rename command 'convert' to 'conserver.convert' to avoid collision with
  a command of same name from imagemagick.
* The 'make test' is renamed to 'make check'.  Later is part of autotools
  standard invocation.
* Add --with-chat and --with-autologin options to build these command.  Both
  are disabled by default.  When enabled these commands are added to
  installation with manual pages.
* Tests work when performing out-of-tree build.  That said distribution
  packages should be created with 'make distcheck'.
* Distribution package is compressed with 'xz', and 'gzip' is omitted.
* The script build-aux/git-version-gen is copied from gnulib to this
  project.  Aim is to make version numbers to be derived from git tag name,
  but this change will need to wait later.
* Add .gitignore file.

All changes to source files are due technical necessity.  These sorts of
changes are kept absolutely minimal in this change, that aims to be as pure
build system update as possible.

But there are few changes, such as in chat.c function logf() had to be
renamed because glibc provides a logarithmic function of same name.  In
autologin.c setpgrp() used arguments, that is rather strange because even
posix defines this function to be used without them.  And project specific
include files are defined quotes rather than angle brackets to give visual
hint what is coming from where.

Reviewed-by: Lyude <thatslyude@gmail.com>