move conserver/version.h values to config.h
Internal major, minor, and revision numbers are removed.  They are not
needed.  Version printing in conserver and console commands can print the
version supplied by git-version-gen that also includes information about
development version, if that is the one that is built.

In same go the VERSION_UINT is represented in config.h as value it is,
rather than in symbolic form it was previously.  Notice that VERSION_UINT
will not include information about development version.